src/main/java/com/sharashkina_kontora/travel_agency/service/UserServiceImpl.java
package com.sharashkina_kontora.travel_agency.service;

import com.sharashkina_kontora.travel_agency.domain.User;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@Service
public class UserServiceImpl implements UserService {
    @Override
    public List<User> findAll() {
        return null;
    }

    @Override
    public Optional<User> findById(Long id) {
        return Optional.empty();
    }

    @Override
    public User save(User user) {
        return null;
    }

    @Override
    public void delete(User user) {

    }
}
